+++ library/string/strcmp.c
@@ -1,7 +1,7 @@
 #include <string.h>
 
 // compares two strings lexicographically
-int strcmp(const char *str1, const char *str2)
+int8_t strcmp(const char *str1, const char *str2)
 {
 	while (*str1 == *str2++) {
 		if (*str1++ == 0) return 0;
